+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * `theme_ggeffects()`, which can be used with `ggplot2::theme_set()` to set the **ggeffects**-theme as default plotting theme. This makes it easier to add further theme-modifications like `sjPlot::legend_style()` or `sjPlot::font_size()`.
 
+## Changes to functions
+
+* The `plot()`-method gets a `ci.style`-argument, to define different styles for the confidence bands for numeric x-axis-terms.
+
 ## Bug fixes
 
 * Fixed issue where confidence intervals could not be computed for variables with very small values, that differ only after the second decimal part.